Fix typo in SvgExtension name
